Move animation tree loading to separate file
