+++ src/game/game.rs
@@ -1,5 +1,3 @@
-use std::collections::HashMap;
-
 use bevy::prelude::*;
 
 use crate::game::components::*;
@@ -105,64 +103,13 @@ fn add_player(
     mut commands : Commands, 
     asset_server: Res<AssetServer>,
     mut texture_atlases: ResMut<Assets<TextureAtlas>>,
-    mut animation_trees : ResMut<Assets<AnimationTree>>) 
+    anim_trees : Res<graphics::AnimationTreeHandles>) 
 {
-    let mut up_transitions = HashMap::new();
-    up_transitions.insert(String::from("down"), String::from("down"));
-    up_transitions.insert(String::from("left"), String::from("left"));
-    up_transitions.insert(String::from("right"), String::from("right"));
-    let up_state = AnimationState {
-        start_frame : 0,
-        end_frame : 3,
-        transitions : up_transitions
-    };
-
-    let mut down_transitions = HashMap::new();
-    down_transitions.insert(String::from("up"), String::from("up"));
-    down_transitions.insert(String::from("left"), String::from("left"));
-    down_transitions.insert(String::from("right"), String::from("right"));
-    let down_state = AnimationState {
-        start_frame : 4,
-        end_frame : 7,
-        transitions : down_transitions
-    };
-
-    let mut left_transitions = HashMap::new();
-    left_transitions.insert(String::from("up"), String::from("up"));
-    left_transitions.insert(String::from("down"), String::from("down"));
-    left_transitions.insert(String::from("right"), String::from("right"));
-    let left_state = AnimationState {
-        start_frame : 8,
-        end_frame : 11,
-        transitions : left_transitions
-    };
-
-    let mut right_transitions = HashMap::new();
-    right_transitions.insert(String::from("up"), String::from("up"));
-    right_transitions.insert(String::from("down"), String::from("down"));
-    right_transitions.insert(String::from("left"), String::from("left"));
-    let right_state = AnimationState {
-        start_frame : 12,
-        end_frame : 15,
-        transitions : right_transitions
-    };
-
-    let mut player_animation_tree = HashMap::new();
-    player_animation_tree.insert(String::from("up"), up_state);
-    player_animation_tree.insert(String::from("down"), down_state);
-    player_animation_tree.insert(String::from("left"), left_state);
-    player_animation_tree.insert(String::from("right"), right_state);
-
     let frame_width : f32 = 10.0;
     let frame_height : f32 = 10.0;
     let texture_handle = asset_server.load("test.png");
     let texture_atlas = TextureAtlas::from_grid(texture_handle, Vec2::new(frame_width, frame_height), 4, 4, Some(Vec2::new(1.0, 1.0)), None);
     let texture_atlas_handle = texture_atlases.add(texture_atlas);
-    let player_anim_tree = AnimationTree {
-        states : player_animation_tree,
-        initial : String::from("right")
-    };
-    let animation_tree_handle = animation_trees.add(player_anim_tree);
     commands.spawn(SpriteSheetBundle {
         texture_atlas : texture_atlas_handle,
         ..default()
@@ -171,7 +118,7 @@ fn add_player(
         timer : Timer::from_seconds(0.1, TimerMode::Repeating),
         events : Vec::new(),
         current_state : String::new(),
-        tree : animation_tree_handle
+        tree : anim_trees.handle_map.get("player").unwrap().clone()
     })
     .insert(RenderInfo {
         screen_width : 100.0,
@@ -213,7 +160,10 @@ enum OrderLabel {
 impl Plugin for Game {
     fn build(&self, app : &mut App) {
         app.init_resource::<input::InputState>()
+            .init_resource::<graphics::AnimationTreeHandles>()
             .add_asset::<AnimationTree>()
+            .add_startup_system(graphics::initialize_anim_trees
+                .before(add_player))
             .add_startup_system(setup_camera)
             .add_startup_system(add_player)
             .add_startup_system(add_enemy)
